pyproject.toml
[tool.poetry]
name = "uniweb-cv-match"
version = "0.1.0"
description = ""
authors = ["kmisterios <kuimov.ms@phystech.edu>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
openai = "^1.59.7"
loguru = "^0.7.3"
numpy = "^2.2.1"
streamlit = "^1.41.1"
pandas = "^2.2.3"
pre-commit = "^4.0.1"
python-dotenv = "^1.0.1"
pyyaml = "^6.0.2"
scikit-learn = "^1.6.1"
tenacity = "^9.0.0"
toml = "^0.10.2"
tqdm = "^4.67.1"
watchdog = "^6.0.0"
nltk = "^3.9.1"
simplemma = "^1.1.2"
rank-bm25 = "^0.2.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
